requirements.txt
langchain
streamlit
gemini
langchain-community
transformers
huggingface-hub
python-dotenv
scipy
